+++ account/transport/public/http/router.go
@@ -1,23 +1,13 @@
 package http
 
 import (
-	"context"
 	"fmt"
 	"github.com/There-is-Go-alternative/GoMicroServices/account/internal/config"
 	"github.com/There-is-Go-alternative/GoMicroServices/account/usecase"
-	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	netHTTP "net/http"
 )
 
-type Server struct {
-	Engine *netHTTP.Server
-	logger zerolog.Logger
-}
-
-type useCase interface {
+type AccountUseCase interface {
 	CreateAccount() usecase.CreateAccountCmd
 	GetAccountByID() usecase.GetAccountByIDCmd
 	GetAllAccounts() usecase.GetAllAccountsCmd
@@ -27,15 +17,7 @@ type useCase interface {
 }
 
 // TODO: change database by future Database interface
-func NewHttpServer(uc useCase, conf *config.Config) *Server {
-	router := gin.Default()
-
-	// Configuring CORS
-	router.Use(cors.Default())
-
-	router.GET("/health", func(c *gin.Context) {
-		c.Status(netHTTP.StatusOK)
-	})
+func ApplyAccountRoutes(router *gin.Engine, uc AccountUseCase, conf *config.Config) {
 	accountHandler := NewAccountHandler()
 	// Grouping Account routes with url specified in config (I.E: 'account')
 	account := router.Group(fmt.Sprintf("/%s", conf.AccountEndpoint))
@@ -47,29 +29,4 @@ func NewHttpServer(uc useCase, conf *config.Config) *Server {
 		account.PUT("/:id", accountHandler.PutAccountHandler(uc.UpdateAccount()))
 		account.DELETE("/:id", accountHandler.DeleteAccountHandler(uc.DeleteAccount()))
 	}
-	return &Server{
-		Engine: &netHTTP.Server{
-			Addr:    fmt.Sprintf("%s:%s", conf.Host, conf.Port),
-			Handler: router,
-		},
-		logger: log.With().Str("service", "HTTP gin server").Logger(),
-	}
-}
-
-func (s Server) Run(ctx context.Context) (err error) {
-	s.logger.Info().Msg("Running gin HTTP server ...")
-	errc := make(chan error)
-	go func() {
-		errc <- s.Engine.ListenAndServe()
-	}()
-	select {
-	case err = <-errc:
-		return
-	case <-ctx.Done():
-		if err = s.Engine.Shutdown(ctx); err != nil && err != context.Canceled {
-			s.logger.Error().Msgf("Error happened when server forced to shutdown: %v", err)
-			return
-		}
-	}
-	return
 }
